requirements.txt
litellm==1.77.5
psutil==7.0.0
python-dotenv==1.1.1
rich==14.1.0
colorama==0.4.6
setproctitle==1.3.3
tqdm==4.66.5
hydra-core==1.3.2
matplotlib==3.10.6
networkx==3.3
numpy==1.24.4
omegaconf==2.3.0
scikit_learn==1.4.2
